Decoupled postgres and command functions from index.js.
